Cut drakrun into pieces, this is my last resort
