+++ drakrun/drakrun/__main__.py
@@ -1,0 +1,3 @@
+from .drakrun import Drakrun
+
+Drakrun.main()
