+++ drakrun/drakrun/machinery/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module contains everything that is needed to set up a VM
+and provides Python interface to OS, Xen and Drakvuf tools
+"""
