+++ drakrun/drakrun/machinery/networking.py
@@ -4,7 +4,7 @@ import signal
 import subprocess
 from typing import Optional
 
-from drakrun.util import get_domid_from_instance_id
+from .vm import get_domid_from_instance_id
 
 log = logging.getLogger("drakrun")
