+++ drakrun/drakrun/profile/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module contains interface to manage VM profiles
+built for snapshot and needed for proper VM introspection
+"""
