--- drakrun/drakrun/py-scripts/drakpdb
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-from drakrun.drakpdb import main
-
-main()
