--- drakrun/drakrun/py-scripts/drakplayground
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-import drakrun.playground as playground
-
-playground.main()
