--- drakrun/drakrun/py-scripts/drakpush
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-
-import drakrun.drakpush as dp
-
-dp.main()
-
