--- drakrun/drakrun/py-scripts/draksetup
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-import drakrun.draksetup as ds
-
-ds.main()
