--- drakrun/drakrun/py-scripts/draktest
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-import drakrun.regression as draktestd
-
-draktestd.RegressionTester.submit_main()
