--- drakrun/drakrun/py-scripts/draktestd
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-import drakrun.regression as draktestd
-
-draktestd.RegressionTester.main()
