+++ drakrun/drakrun/templates/config.dist.ini
@@ -4,12 +4,11 @@
 host=localhost
 port=6379
 
-[minio]
+[s3]
 ; MinIO server
 ; used to store job queue and analysis results
-address=localhost:9000
+address=http://localhost:9000
 bucket=karton
-secure=0
 
 ; MinIO access credentials
 ;
