+++ drakrun/drakrun/templates/hooks.dist.txt
