+++ drakrun/drakrun/test/test_network.py
@@ -5,8 +5,8 @@ from pathlib import Path
 import pytest
 from pytest_steps import depends_on, test_steps
 
-from drakrun.draksetup import find_default_interface
-from drakrun.networking import (
+from drakrun.utilities.draksetup import find_default_interface
+from drakrun.machinery.networking import (
     add_iptable_rule,
     del_iptable_rule,
     delete_vm_network,
