+++ drakrun/drakrun/test/test_vm.py
@@ -10,7 +10,7 @@ from common_utils import remove_files, tool_exists
 
 from drakrun.config import InstallInfo
 from drakrun.util import safe_delete
-from drakrun.vm import VirtualMachine
+from drakrun.machinery.vm import VirtualMachine
 
 
 @pytest.fixture(scope="session")
