+++ drakrun/drakrun/utilities/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module contains CLI utilities that are needed
+to set up and test Drakvuf Sandbox installation
+"""
